move App.tsx out of src/components
